Edit event (#80)
* reschedule and edit events

* delete ac

* Update scheduled events.

Moved rescheduleScheduledEvent functionality from scheduledeventcontroller to scheduledeventserver. Now not only the start date of an event can be edited, but also the end date, courses and scenarios, etc.

* Delete scheduled events

* namespace in labels, deleteCollection instead of list and delete

* Delete VMs

* Finish sessions.
Now, when a scheduled event is deleted, the associated sessions to the event's courses/scenarios are terminated. Terminating a session also deletes the corresponding VMs.

* invalidate sessions for finished events

* Delete DBRs when cleaning up session

* Change Label for DBRs to Session instead of VMClaim

Co-authored-by: Jan-Gerrit Goebel <Jan-Gerrit.Goebel@sva.de>